Split minmaxany
