+++ src/AggregateFunctions/AggregateFunctionSimpleState.h
@@ -71,7 +71,7 @@ public:
 
     bool allocatesMemoryInArena() const override { return nested_func->allocatesMemoryInArena(); }
 
-    AggregateFunctionPtr getNestedFunction() const { return nested_func; }
+    AggregateFunctionPtr getNestedFunction() const override { return nested_func; }
 };
 
 }
